+++ src/fintrist_ds/catalog.py
@@ -42,14 +42,17 @@ def backtest(model, strategy, period='1y', end=None):
     simulated = []
     tempstudy = Study()
     parent_data = {name: study.data for name, study in model.parents.items()}
-    function = CATALOG[model.process.name]
+    try:
+        function = ANALYSIS_CATALOG[model.process.name]
+    except KeyError:
+        function = SCRAPERS_CATALOG[model.process.name]
+        parent_data['mock'] = model.data
 
     # At each date, run the model's function on the previous data
     # TODO: Date range should be based on the model's parents, not the model.
     full_range = model.data.index
     for view_date in model.data[start.date():end.date()].index:
-        print(f"Backtesting at {view_date}")
-        logger.info(f"Log: Backtesting at {view_date}")
+        logger.debug(f"Log: Backtesting at {view_date}")
         curr_idx = full_range.get_loc(view_date)
         try:
             prev_date = full_range[curr_idx - 1]