+++ src/fintrist_ds/scrapers/base.py
@@ -7,13 +7,17 @@ from fintrist_ds.settings import Config
 
 __all__ = ['stock']
 
-def stock(symbol, frequency='daily', source=None):
+def stock(symbol, frequency='daily', source=None, mock=None):
     """Get a stock quote history.
 
+    ::parents:: mock
     ::params:: symbol, frequency, source
     ::alerts:: source: AV, source: Tiingo, ex-dividend, split, reverse split
     """
-    if not source:
+    ## Get the data from whichever source
+    if mock is not None:
+        source = 'mock'
+    elif not source:
         source = 'Tiingo'
     if source == 'AV':
         data = pdr.get_data_alphavantage(symbol, api_key=Config.APIKEY_AV, start='1900')
@@ -22,6 +26,8 @@ def stock(symbol, frequency='daily', source=None):
         data = pdr.get_data_tiingo(symbol, api_key=Config.APIKEY_TIINGO, start='1900')
         data = data.droplevel('symbol')  # Multiple stock symbols are possible
         data.index = data.index.date
+    elif source == 'mock':
+        data = mock
 
     ## Create alerts
     alerts = [f'source: {source}']